Fixing tests after refactor
