+++ front/src/components/__tests__/ToolBox/Actions/Action.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Action from '../../actions/Action';
+import Action from '../../../ToolBox/Actions/Action';
 
 describe('Action component', () => {
   it('it can not be instantiated', () => {
