+++ front/src/components/__tests__/ToolBox/Machines/Crafter.spec.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import Seller from '../../machines/Seller';
+import Crafter from '../../../ToolBox/Machines/Crafter';
 
-describe('Seller Component', () => {
+describe('Crafter Component', () => {
   it('has an img tag', () => {
-    const component = shallow(<Seller />);
+    const component = shallow(<Crafter />);
     const node = component.find('img');
     expect(node.length).toEqual(1);
   });
 
   it('has an machine className', () => {
-    const component = mount(<Seller />);
+    const component = mount(<Crafter />);
     expect(component.find('img').hasClass('machine')).toBeTruthy();
   });
 });
