+++ front/src/components/__tests__/ToolBox/Machines/Machine.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Machine from '../../machines/Machine';
+import Machine from '../../../ToolBox/Machines/Machine';
 
 describe('Machine component', () => {
   it('it can not be instantiated', () => {
