+++ front/src/components/__tests__/ToolBox/Machines/Starter.spec.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import Transporter from '../../machines/Transporter';
+import Starter from '../../../ToolBox/Machines/Starter';
 
-describe('Transporter Component', () => {
+describe('Starter Component', () => {
   it('has an img tag', () => {
-    const component = shallow(<Transporter />);
+    const component = shallow(<Starter />);
     const node = component.find('img');
     expect(node.length).toEqual(1);
   });
 
   it('has an machine className', () => {
-    const component = mount(<Transporter />);
+    const component = mount(<Starter />);
     expect(component.find('img').hasClass('machine')).toBeTruthy();
   });
 });
