+++ front/src/components/__tests__/ToolBox/Machines/Transporter.spec.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import Crafter from '../../machines/Crafter';
+import Transporter from '../../../ToolBox/Machines/Transporter';
 
-describe('Crafter Component', () => {
+describe('Transporter Component', () => {
   it('has an img tag', () => {
-    const component = shallow(<Crafter />);
+    const component = shallow(<Transporter />);
     const node = component.find('img');
     expect(node.length).toEqual(1);
   });
 
   it('has an machine className', () => {
-    const component = mount(<Crafter />);
+    const component = mount(<Transporter />);
     expect(component.find('img').hasClass('machine')).toBeTruthy();
   });
 });
