+++ front/src/components/__tests__/ToolBox/ToolboxComponent.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import ToolboxComponent from '../ToolboxComponent';
+import ToolboxComponent from '../../ToolBox/ToolboxComponent';
 
 describe('ToolboxComponent', () => {
   it('it can not be instantiated', () => {
